+++ src/parser/smt2/parser.cpp
@@ -17,8 +17,10 @@ namespace parser::smt2 {
 
 /* Parser public ------------------------------------------------------------ */
 
-Parser::Parser(bitwuzla::Options& options, const std::string& infile_name)
-    : bzla::parser::Parser(options, infile_name)
+Parser::Parser(bitwuzla::Options& options,
+               const std::string& infile_name,
+               std::ostream* out)
+    : bzla::parser::Parser(options, infile_name, out)
 {
   if (d_error.empty())
   {
@@ -38,8 +40,9 @@ Parser::Parser(bitwuzla::Options& options, const std::string& infile_name)
 
 Parser::Parser(bitwuzla::Options& options,
                const std::string& infile_name,
-               FILE* infile)
-    : bzla::parser::Parser(options, infile_name, infile)
+               FILE* infile,
+               std::ostream* out)
+    : bzla::parser::Parser(options, infile_name, infile, out)
 {
   if (d_error.empty())
   {